backend/requirements.txt
# Core Flask dependencies
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-Limiter==3.5.0

# Database
psycopg2-binary==2.9.7
PyMySQL==1.1.0
SQLAlchemy==2.0.21

# Payment Gateway (Midtrans)

# Selenium for automation
selenium==4.15.2
webdriver-manager==4.0.1

# Async tasks
celery==5.3.4
redis==5.0.1

# Email services
sendgrid==6.10.0

# Security & Validation
bcrypt==4.0.1
email-validator==2.0.0
phonenumbers==8.13.23

# Environment & Configuration
python-dotenv==1.0.0

# Utilities
requests==2.31.0
Pillow==10.0.1
python-dateutil==2.8.2

# Development
pytest==7.4.2
pytest-flask==1.3.0